media-service/kafka-consumer/requirement.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==5.0.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
PyYAML==6.0.2
Werkzeug==3.0.4
zipp==3.20.2
minio
kafka-python
redis
psycopg2-binary
boto3
pillow